Update tests for all chechout page components
